Move classes around now that we no longer have inheritance.
